Update README to reflect dependencies
A clean clone of the repo won't build, since it depends on https://github.com/soffes/x. 

I'm not really a Carthage person, so I can't speak to whether the actual correct solution is to check in dependencies or not, but for now updating the README seems like a sensible quick fix.